refactor(settings): Move cards and components inside their corresponding routes for better management
