+++ apps/web/src/app/[locale]/(dashboard)/settings/account/billing/page.tsx
@@ -1,4 +1,5 @@
-import { CurrentPlanCard, PaymentBillingCard } from '../components';
+import CurrentPlanCard from './current-plan-card';
+import PaymentBillingCard from './payment-billing-card';
 
 export default async function BillingPage() {
   return (
