--- apps/web/src/app/[locale]/(dashboard)/settings/account/components/index.ts
@@ -1,11 +1,0 @@
-export { default as AccountStatusSection } from './account-status-section';
-export { default as ContactInformationCard } from './contact-information-card';
-export { default as CurrentPlanCard } from './current-plan-card';
-export { default as NotificationsCard } from './notifications-card';
-export { default as PageHeader } from './page-header';
-export { default as PaymentBillingCard } from './payment-billing-card';
-export { default as ProfileInformationCard } from './profile-information-card';
-export { default as SecuritySettingsCard } from './security-settings-card';
-export { default as SessionsCard } from './sessions-card';
-export { default as SettingsNav } from './settings-nav';
-export { default as WorkspaceSettingsCard } from './workspace-settings-card';
