+++ apps/web/src/app/[locale]/(dashboard)/settings/account/notifications/notifications-card.tsx
