+++ apps/web/src/app/[locale]/(dashboard)/settings/account/notifications/page.tsx
@@ -1,4 +1,4 @@
-import { NotificationsCard } from '../components';
+import NotificationsCard from './notifications-card';
 
 export default async function NotificationsPage() {
   return <NotificationsCard />;
