+++ apps/web/src/app/[locale]/(dashboard)/settings/account/page.tsx
@@ -1,4 +1,5 @@
-import { ContactInformationCard, ProfileInformationCard } from './components';
+import ContactInformationCard from './contact-information-card';
+import ProfileInformationCard from './profile-information-card';
 import { getCurrentUser } from '@tuturuuu/utils/user-helper';
 
 export default async function AccountSettingsPage() {
