+++ apps/web/src/app/[locale]/(dashboard)/settings/account/security/mfa-card.tsx
@@ -1,0 +1,3 @@
+export default function MFACard() {
+  return <div>MFACard</div>;
+}
