+++ apps/web/src/app/[locale]/(dashboard)/settings/account/security/reset-password-dialog.tsx
