+++ apps/web/src/app/[locale]/(dashboard)/settings/account/security/security-settings-card.tsx
@@ -1,4 +1,4 @@
-import ResetPasswordDialog from '../reset-password-dialog';
+import ResetPasswordDialog from './reset-password-dialog';
 import type { WorkspaceUser } from '@tuturuuu/types/primitives/WorkspaceUser';
 import {
   Card,
