+++ apps/web/src/app/[locale]/(dashboard)/settings/account/sessions/page.tsx
@@ -1,4 +1,4 @@
-import { SessionsCard } from '../components';
+import SessionsCard from './sessions-card';
 
 export default async function SessionsPage() {
   return <SessionsCard />;
