+++ apps/web/src/app/[locale]/(dashboard)/settings/account/sessions/sessions-card.tsx
