+++ apps/web/src/app/[locale]/(dashboard)/settings/account/workspaces/default-workspace-setting.tsx
